requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos>=1.2.0
apache-airflow-providers-snowflake>=4.0.0,<6